backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
httpx>=0.24.0
Pillow>=10.0.0
reportlab>=4.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0